working version after refactoring
